refactor: 문제 설정 선택 컴포넌트 분리
