+++ src/pages/Main/TestMake/TestMake.tsx
@@ -1,9 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import styles from './TestMake.module.scss';
 import { getAllScripts, IScriptInList } from '../../../apis/script';
-import QuestionTypeBtn from '../../../components/common/buttons/QuestionTypeBtn/QuestionTypeBtn';
-import useTestSettingsStore from '../../../store/TestSettingsStore';
 import { useNavigate } from 'react-router-dom';
+import TestOptionSelector from '../../../components/TestOptionSelector/TestOptionSelector';
 
 const TestMake = () => {
   const navigate = useNavigate();
@@ -12,26 +11,6 @@ const TestMake = () => {
     queryKey: ['allScripts'],
     queryFn: () => getAllScripts(),
   });
-
-  const { questionCount, timeLimit, setQuestionCount, setTimeLimit } =
-    useTestSettingsStore();
-
-  const handleQuestionCountChange = (
-    e: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    setQuestionCount(parseInt(e.target.value, 10));
-  };
-
-  const handleTimeLimitChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const checked = e.target.checked;
-    setTimeLimit(checked ? 0 : null);
-  };
-
-  const handleTimeLimitValueChange = (
-    e: React.ChangeEvent<HTMLInputElement>,
-  ) => {
-    setTimeLimit(parseInt(e.target.value, 10));
-  };
 
   const handleTestStartBtnClick = () => {
     // 문제 유형 유효성 검사 로직 구현 예정
@@ -60,55 +39,7 @@ const TestMake = () => {
             </p>
           </section>
         )}
-        <section className={styles.testOptionContainer}>
-          <div className={styles.selectBox}>
-            <p className={styles.selectTitle}>문제 유형</p>
-            <div className={styles.selectionBtnsContainer}>
-              {['객관식', '단답형', '빈칸 뚫기', 'OX 퀴즈'].map(
-                (type, index) => (
-                  <QuestionTypeBtn key={index}>{type}</QuestionTypeBtn>
-                ),
-              )}
-            </div>
-          </div>
-          <div className={styles.selectBox}>
-            <p className={styles.selectTitle}>문제 개수</p>
-            <input
-              className={styles.numInput}
-              type="number"
-              min="0"
-              value={questionCount}
-              onChange={handleQuestionCountChange}
-            />
-            <span className={styles.inputCaption}>개</span>
-          </div>
-          <div className={styles.selectBox}>
-            <span className={styles.checkBoxContainer}>
-              <input
-                className={styles.timeLimitInput}
-                type="checkbox"
-                id="time"
-                checked={timeLimit !== null}
-                onChange={handleTimeLimitChange}
-              />
-              <label className={styles.checkBoxTitle} htmlFor="time">
-                시간 제한
-              </label>
-            </span>
-            {timeLimit !== null && (
-              <>
-                <input
-                  className={styles.numInput}
-                  type="number"
-                  min="0"
-                  value={timeLimit}
-                  onChange={handleTimeLimitValueChange}
-                />
-                <span className={styles.inputCaption}>분</span>
-              </>
-            )}
-          </div>
-        </section>
+        <TestOptionSelector />
       </div>
     </div>
   );
